For testing new refactored code
