+++ run_fps.py
@@ -1,5 +1,5 @@
 from multiprocessing import Pool
-from lite_deepsort_app import run
+from track import run
 from opts import opt
 import time
 from os.path import join
@@ -15,7 +15,6 @@ def process_sequence(seq):
     run(
         sequence_dir=join(opt.dir_dataset, seq),
         output_file=path_save,
-        min_confidence=opt.min_confidence,
         nn_budget=opt.nn_budget,
         display=False,
         visualize=False,
