+++ solutions.py
