+++ trackers/__init__.py
@@ -1,0 +1,6 @@
+from .strongsort import StrongSORT
+from .lite import LITE
+from .deepsort import DeepSORT
+from .osnet import OSNet
+
+__all__ = "StrongSORT", "LITE", "DeepSORT", "OSNet"
